docs(information-architecture): update navigation (#926)
* docs(information-architecture): update navigation

* chore: merge `dev` into `layout-module-information-architecture`
