+++ .storybook/index.js
@@ -1,19 +1,35 @@
 /**
  * @file Storybook helpers.
- * @copyright IBM Security 2019 - 2020
+ * @copyright IBM Security 2019 - 2021
  */
+
+import { OVERVIEW } from '../src/components/UNSTABLE__LayoutModules/stories';
 
 // Category labels.
 const CATEGORIES = {
   COMPONENTS: 'Components',
+  DEPRECATED: 'Deprecated',
+  LAYOUT_MODULES: 'Layout modules',
+  PAGE_LAYOUTS: 'Page layouts (Canary)',
   PATTERNS: 'Patterns',
 };
 
-const { COMPONENTS, PATTERNS } = CATEGORIES;
+const {
+  COMPONENTS,
+  DEPRECATED,
+  LAYOUT_MODULES,
+  PAGE_LAYOUTS,
+  PATTERNS,
+} = CATEGORIES;
 
-const HIERARCHY_ROOT_SEPARATOR = '/';
-
-const ORDER = [PATTERNS, COMPONENTS];
+const ORDER = [
+  LAYOUT_MODULES,
+  [OVERVIEW],
+  PAGE_LAYOUTS,
+  PATTERNS,
+  COMPONENTS,
+  DEPRECATED,
+];
 
 /**
  * Returns a formatted string for the Storybook category.
@@ -21,8 +37,7 @@ const ORDER = [PATTERNS, COMPONENTS];
  * @param {string} storyName The story name to format.
  * @returns {string} The formatted category and story name.
  */
-const getCategory = (categoryName, storyName) =>
-  `${categoryName}${HIERARCHY_ROOT_SEPARATOR}${storyName}`;
+const getCategory = (categoryName, storyName) => `${categoryName}/${storyName}`;
 
 /**
  * Binds Storybook category to the `getCategory` method.
@@ -38,18 +53,32 @@ const bindCategory = categoryName => getCategory.bind(this, categoryName);
 const getComponentsCategory = bindCategory(COMPONENTS);
 
 /**
+ * Returns a formatted string for the deprecated category.
+ * @param {string} storyName The story name to format.
+ * @returns {string} The formatted deprecated category and story name.
+ */
+const getDeprecatedCategory = bindCategory(DEPRECATED);
+
+/**
+ * Returns a formatted string for the layout modules category.
+ * @param {string} storyName The story name to format.
+ * @returns {string} The formatted layout module category and story name.
+ */
+const getLayoutModulesCategory = bindCategory(LAYOUT_MODULES);
+
+/**
+ * Returns a formatted string for the page layouts category.
+ * @param {string} storyName The story name to format.
+ * @returns {string} The formatted page layout category and story name.
+ */
+const getPageLayoutsCategory = bindCategory(PAGE_LAYOUTS);
+
+/**
  * Returns a formatted string for the patterns category.
  * @param {string} storyName The story name to format.
  * @returns {string} The formatted pattern category and story name.
  */
 const getPatternsCategory = bindCategory(PATTERNS);
-
-/**
- * Returns a formatted string for deprecated components.
- * @param {string} name The name to deprecated.
- * @returns {string} The formatted deprecated string.
- */
-const deprecate = name => `${name} [Deprecated]`;
 
 /**
  * Configuration for disabling the centering addon.
@@ -101,13 +130,14 @@ Also refer to http://${component.library}.carbondesignsystem.com/?path=/story/${
 });
 
 export {
-  deprecate,
   disableCentered,
   disableCenteredStories,
   getComponentsCategory as components,
+  getDeprecatedCategory as deprecated,
   getDocsParameters,
+  getLayoutModulesCategory as layoutModules,
+  getPageLayoutsCategory as pageLayouts,
   getPatternsCategory as patterns,
-  HIERARCHY_ROOT_SEPARATOR,
   info,
   ORDER,
 };
