+++ src/components/UNSTABLE__LayoutModules/ActionBarModule/index.mdx
@@ -25,10 +25,7 @@ The action bar module can be built using a combination of the `ActionBarModule` 
 `ActionBarModuleItems` is responsible for positioning the actions, and everything passed in as a child of `ActionBarModuleItems` will be rendered.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-actionbarmodule--default"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-actionbarmodule--default" parameters={parameters} />
 </Canvas>
 
 ### Component API
@@ -52,10 +49,7 @@ Additional props passed into `ActionBarModule` and `ActionBarModuleItems` will b
 If necessary, supplementary details can be placed inline with the actions instead of the ghost button.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-actionbarmodule--variant"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-actionbarmodule--variant" parameters={parameters} />
 </Canvas>
 
 ### Grid placement
