+++ src/components/UNSTABLE__LayoutModules/ButtonClusterModule/index.mdx
@@ -26,7 +26,7 @@ The button cluster module can be built using a combination of the `ButtonCluster
 
 <Canvas>
   <Story
-    id="patterns-unstable-layout-modules-buttonclustermodule--default"
+    id="layout-modules-buttonclustermodule--default"
     parameters={parameters}
   />
 </Canvas>
