+++ src/components/UNSTABLE__LayoutModules/CardModule/index.mdx
@@ -27,10 +27,7 @@ The card module can be built using a combination of the `CardModule`, [grid](htt
 Everything passed in as a [render prop](https://reactjs.org/docs/render-props.html) child of `CardModule` will be rendered.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-cardmodule--default"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-cardmodule--default" parameters={parameters} />
 </Canvas>
 
 ### Component API
@@ -68,7 +65,7 @@ The cards within the module can also be arranged in a number of different ways.
 
 <img alt="`CardModule` variants" src={variants} />
 
-<Source id="patterns-unstable-layout-modules-cardmodule--variant" />
+<Source id="layout-modules-cardmodule--variant" />
 
 ### Grid placement
 
@@ -107,16 +104,10 @@ Card modules can span between 4 and 16 columns. Following the narrow grid implem
   </LinkTo>
 - [Tile](https://react.carbondesignsystem.com/?path=/docs/tile)
 - [Render Props](https://reactjs.org/docs/render-props.html)
-- <LinkTo
-    kind="patterns-unstable-layout-modules-titlebarmodule"
-    story="default"
-  >
+- <LinkTo kind="layout-modules-titlebarmodule" story="default">
     TitleBarModule
   </LinkTo>
-- <LinkTo
-    kind="patterns-unstable-layout-modules-actionbarmodule"
-    story="default"
-  >
+- <LinkTo kind="layout-modules-actionbarmodule" story="default">
     ActionBarModule
   </LinkTo>
 - <LinkTo kind="patterns-iconbuttonbar" story="default">
