+++ src/components/UNSTABLE__LayoutModules/DescriptionListModule/index.mdx
@@ -27,13 +27,13 @@ Description list modules provide a means to orderly layout terms and definitions
 
 ### Overview
 
-The description list module can be built using a combination of the <LinkTo kind="components-typelayout" story="default">`TypeLayout`</LinkTo> and <LinkTo kind="patterns-unstable-layout-modules-titlebarmodule" story="default">`TitleBarModule`</LinkTo> components.
+The description list module can be built using a combination of the <LinkTo kind="components-typelayout" story="default">`TypeLayout`</LinkTo> and <LinkTo kind="layout-modules-titlebarmodule" story="default">`TitleBarModule`</LinkTo> components.
 
 Everything passed in as a child of `DescriptionListModule` will be rendered.
 
 <Canvas>
   <Story
-    id="patterns-unstable-layout-modules-descriptionlistmodule--default"
+    id="layout-modules-descriptionlistmodule--default"
     parameters={parameters}
   />
 </Canvas>
@@ -57,11 +57,11 @@ Additional props passed into `DescriptionListModule` will be forwarded along to 
 
 ### Variants
 
-If necessary, both section and subsection titles can precede the type layout. There are two values already supplied for this so you shouldn't supplement your own type style. Learn more by reviewing the <LinkTo kind="patterns-unstable-layout-modules-titlebarmodule" story="default" > TitleBarModule. </LinkTo>
+If necessary, both section and subsection titles can precede the type layout. There are two values already supplied for this so you shouldn't supplement your own type style. Learn more by reviewing the <LinkTo kind="layout-modules-titlebarmodule" story="default" > TitleBarModule. </LinkTo>
 
 <img alt="`DescriptionListModule` variants" src={variants} />
 
-<Source id="patterns-unstable-layout-modules-descriptionlistmodule--variant" />
+<Source id="layout-modules-descriptionlistmodule--variant" />
 
 ### Grid placement
 
@@ -106,10 +106,7 @@ Occassionally a [spacing token](https://carbondesignsystem.com/guidelines/spacin
 - <LinkTo kind="components-typelayout" story="default">
     TypeLayout
   </LinkTo>
-- <LinkTo
-    kind="patterns-unstable-layout-modules-titlebarmodule"
-    story="default"
-  >
+- <LinkTo kind="layout-modules-titlebarmodule" story="default">
     TitleBarModule
   </LinkTo>
 - [Grid](https://react.carbondesignsystem.com/?path=/docs/grid--narrow)
