+++ src/components/UNSTABLE__LayoutModules/DescriptionModule/index.mdx
@@ -25,13 +25,13 @@ Description modules provide a means to orderly layout short-form content.
 
 ### Overview
 
-The description module can be built using a combination of the <LinkTo kind="patterns-unstable-layout-modules-titlebarmodule" story="default">`TitleBarModule`</LinkTo> component and [paragraph elements](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/p).
+The description module can be built using a combination of the <LinkTo kind="layout-modules-titlebarmodule" story="default">`TitleBarModule`</LinkTo> component and [paragraph elements](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/p).
 
 Everything passed in as a [render prop](https://reactjs.org/docs/render-props.html) child of `DescriptionModule` will be rendered.
 
 <Canvas>
   <Story
-    id="patterns-unstable-layout-modules-descriptionmodule--default"
+    id="layout-modules-descriptionmodule--default"
     parameters={parameters}
   />
 </Canvas>
@@ -54,7 +54,7 @@ In order to make sure that everything works as intended, it's important to pass 
 
 ### Anatomy
 
-1. **Title:** This is the title, achieved by using the `TitleBarModule`. Learn more about the <LinkTo kind="patterns-unstable-layout-modules-titlebarmodule" story="default">TitleBarModule</LinkTo> and potential variants.
+1. **Title:** This is the title, achieved by using the `TitleBarModule`. Learn more about the <LinkTo kind="layout-modules-titlebarmodule" story="default">TitleBarModule</LinkTo> and potential variants.
 2. **Description:** This is the description itself, the type style is defined so you shouldn't supplement your own values.
 3. **Bottom spacer:** There is a 32-pixel `$layout-03` spacer included below the description. Any additional modules can be placed directly underneath.
 
@@ -62,13 +62,13 @@ In order to make sure that everything works as intended, it's important to pass 
 
 ### Variants
 
-If necessary, subsection titles can be substituted or used below section titles. There are values already supplied for the title type styles so you shouldn't supplement your own. Learn more by reviewing the <LinkTo kind="patterns-unstable-layout-modules-titlebarmodule" story="default" > TitleBarModule. </LinkTo>
+If necessary, subsection titles can be substituted or used below section titles. There are values already supplied for the title type styles so you shouldn't supplement your own. Learn more by reviewing the <LinkTo kind="layout-modules-titlebarmodule" story="default" > TitleBarModule. </LinkTo>
 
 The description module can also be created without any title.
 
 <img alt="`DescriptionModule` variants" src={variants} />
 
-<Source id="patterns-unstable-layout-modules-descriptionmodule--variant" />
+<Source id="layout-modules-descriptionmodule--variant" />
 
 ### Grid placement
 
@@ -107,10 +107,7 @@ Occassionally a [spacing token](https://carbondesignsystem.com/guidelines/spacin
 
 ### References
 
-- <LinkTo
-    kind="patterns-unstable-layout-modules-titlebarmodule"
-    story="default"
-  >
+- <LinkTo kind="layout-modules-titlebarmodule" story="default">
     TitleBarModule
   </LinkTo>
 - [The Paragraph element](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/p)
