+++ src/components/UNSTABLE__LayoutModules/ICAModule/index.mdx
@@ -28,10 +28,7 @@ The ICA module can be built using a combination of the `ICAModule`, [grid](https
 Everything passed in as a [render prop](https://reactjs.org/docs/render-props.html) child of `ICAModule` will be rendered.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-icamodule--default"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-icamodule--default" parameters={parameters} />
 </Canvas>
 
 ### Component API
@@ -76,7 +73,7 @@ If the ICA needs to be interactive, a background hover can be placed on the grid
 
 <img alt="`ICAModule` interactions" src={interactions} />
 
-<Source id="patterns-unstable-layout-modules-icamodule--interaction" />
+<Source id="layout-modules-icamodule--interaction" />
 
 ### Grid placement
 
@@ -126,10 +123,7 @@ ICAs within ICA modules can span across 2 or 3 columns, following the narrow gri
     ICA
   </LinkTo>
 - [Render Props](https://reactjs.org/docs/render-props.html)
-- <LinkTo
-    kind="patterns-unstable-layout-modules-titlebarmodule"
-    story="default"
-  >
+- <LinkTo kind="layout-modules-titlebarmodule" story="default">
     TitleBarModule
   </LinkTo>
