+++ src/components/UNSTABLE__LayoutModules/PageTabModule/index.mdx
@@ -25,10 +25,7 @@ The page tab module can be built using a combination of the `PageTabModule`, [`T
 The `PageTabModule` accepts `Tabs`, and optionally `PageTabModuleDetails`, components as `children` and each `Tab` is responsible for displaying the tab's label and panel content.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-pagetabmodule--default"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-pagetabmodule--default" parameters={parameters} />
 </Canvas>
 
 ### Variations
@@ -36,10 +33,7 @@ The `PageTabModule` accepts `Tabs`, and optionally `PageTabModuleDetails`, compo
 If necessary, supplementary details can be placed inline with the tabs.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-pagetabmodule--variant"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-pagetabmodule--variant" parameters={parameters} />
 </Canvas>
 
 ### Component API
