+++ src/components/UNSTABLE__LayoutModules/TitleBarModule/index.mdx
@@ -26,10 +26,7 @@ Title bar modules provide interchangeable and reliable headings for establishing
 Everything passed in as a child of the title bar module will be rendered.
 
 <Canvas>
-  <Story
-    id="patterns-unstable-layout-modules-titlebarmodule--default"
-    parameters={parameters}
-  />
+  <Story id="layout-modules-titlebarmodule--default" parameters={parameters} />
 </Canvas>
 
 ### Component API
@@ -80,7 +77,7 @@ The title bar can also accept two type variants, sub-section title and section t
 
 <img alt="`TitleBarModule` variants" src={variants} />
 
-<Source id="patterns-unstable-layout-modules-titlebarmodule--variant" />
+<Source id="layout-modules-titlebarmodule--variant" />
 
 ### Grid placement
