+++ src/components/UNSTABLE__LayoutModules/docs/examples/Detail/index.mdx
@@ -1,11 +1,7 @@
 import detail from './images/detail.png';
 
 <Column lg={8}>
-  <LinkTo
-    className="storybook__link--image"
-    kind="patterns-unstable-modular-page-layouts"
-    story="detail"
-  >
-    <img alt="Detail layout page example" src={detail} />
+  <LinkTo className="storybook__link--image" kind="page-layouts-canary-detail">
+    <img alt="Detail page layout example" src={detail} />
   </LinkTo>
 </Column>
