+++ src/components/UNSTABLE__LayoutModules/docs/examples/Overview/index.mdx
@@ -3,9 +3,8 @@ import overview from './images/overview.png';
 <Column lg={8}>
   <LinkTo
     className="storybook__link--image"
-    kind="patterns-unstable-modular-page-layouts"
-    story="overview"
+    kind="page-layouts-canary-overview"
   >
-    <img alt="Overview layout page example" src={overview} />
+    <img alt="Overview page layout example" src={overview} />
   </LinkTo>
 </Column>
