+++ src/components/UNSTABLE__LayoutModules/stories/Overview.stories.mdx
@@ -1,5 +1,4 @@
 import { Meta } from '@storybook/addon-docs/blocks';
+import getTitle, { OVERVIEW } from '.';
 
-import getTitle from '.';
-
-<Meta title={getTitle('Overview')} />
+<Meta title={getTitle(OVERVIEW)} />
