+++ src/components/UNSTABLE__LayoutModules/stories/index.js
@@ -3,6 +3,8 @@
  * @copyright IBM Security 2020 - 2021
  */
 
-import { patterns } from '../../../../.storybook';
+import { layoutModules } from '../../../../.storybook';
 
-export default title => patterns(`UNSTABLE Layout Modules/${title}`);
+export const OVERVIEW = 'Overview';
+
+export default title => layoutModules(title);
