+++ src/components/UNSTABLE__PageLayouts/Overview.stories.js
@@ -1,40 +1,23 @@
 /**
- * @file Layout module stories.
+ * @file Overview page layout stories.
  * @copyright IBM Security 2020 - 2021
  */
 
-import {
-  Activity16,
-  Copy16,
-  Filter16,
-  Search16,
-  View16,
-} from '@carbon/icons-react';
+import { Activity16, Copy16 } from '@carbon/icons-react';
 
-import { layout03, layout04 } from '@carbon/layout';
-
-import { Grid, Row, Column } from 'carbon-components-react';
+import { Row, Column } from 'carbon-components-react';
 import React from 'react';
 
-import { disableCentered, patterns } from '../../../.storybook';
+import { pageLayouts } from '../../../.storybook';
 
 import {
   ActionBarModule,
-  ActionBarModuleItems,
   ButtonClusterModule,
   Button,
-  CarbonHeader,
-  CardModule,
   DataTablePagination,
-  Decorator,
-  DescriptionModule,
   DescriptionListModule,
-  HeaderName,
   ICA,
   ICAModule,
-  IconButtonBar,
-  SummaryCard as SummaryCardComponent,
-  SummaryCardHeader,
   Tag,
   TitleBarModule,
   TypeLayout,
@@ -44,224 +27,19 @@ import {
   withBackground,
 } from '../..';
 
-const ColumnWithBackground = withBackground(Column);
-const SummaryCard = withBackground(SummaryCardComponent);
+import config from './stories';
 
-const UIShell = () => (
-  <div style={{ height: layout04, marginBottom: layout03 }}>
-    <CarbonHeader aria-label="IBM Security">
-      <HeaderName prefix="IBM">Security</HeaderName>
-    </CarbonHeader>
-  </div>
-);
+const { decorators, parameters } = config;
+
+const ColumnWithBackground = withBackground(Column);
 
 export default {
-  title: patterns('UNSTABLE Modular Page Layouts'),
-  parameters: {
-    ...disableCentered(),
-  },
-  decorators: [
-    story => (
-      <>
-        <UIShell />
-
-        <Grid>{story()}</Grid>
-      </>
-    ),
-  ],
+  title: pageLayouts('Overview'),
+  parameters,
+  decorators,
 };
 
-export const Detail = () => (
-  <>
-    <TitleBarModule title="Summary" />
-
-    <Row>
-      <Column lg={6}>
-        <DescriptionModule>
-          {({ getLayoutProps }) => (
-            <p {...getLayoutProps()}>
-              BadFlick is a backdoor that is usually seen being distributed
-              using exploited word documents. It does not have any persistence
-              to survive reboot, but it is capable of opening a reverse shell
-              connection to its C2 server where it can download and execute
-              possibly other malware.
-            </p>
-          )}
-        </DescriptionModule>
-      </Column>
-
-      <Column lg={{ offset: 2, span: 8 }}>
-        <DescriptionListModule>
-          <TypeLayout>
-            <TypeLayoutBody>
-              <TypeLayoutRow>
-                <TypeLayoutCell>Created by</TypeLayoutCell>
-                <TypeLayoutCell>X-Force IRIS</TypeLayoutCell>
-              </TypeLayoutRow>
-              <TypeLayoutRow>
-                <TypeLayoutCell>Last updated</TypeLayoutCell>
-                <TypeLayoutCell>Jul 14 2019</TypeLayoutCell>
-              </TypeLayoutRow>
-            </TypeLayoutBody>
-          </TypeLayout>
-        </DescriptionListModule>
-      </Column>
-    </Row>
-
-    <TitleBarModule title="Related reports" />
-
-    <ActionBarModule>
-      Supplementary details
-      <ActionBarModuleItems>
-        <IconButtonBar
-          actions={[
-            {
-              label: 'Search',
-              renderIcon: Search16,
-            },
-            {
-              label: 'Filter',
-              renderIcon: Filter16,
-            },
-            {
-              label: 'View',
-              renderIcon: View16,
-            },
-          ]}
-          size="lg"
-        />
-      </ActionBarModuleItems>
-    </ActionBarModule>
-
-    <CardModule>
-      {({ getLayoutProps }) => (
-        <Row>
-          <Column>
-            <SummaryCard {...getLayoutProps()}>
-              <SummaryCardHeader title="Threat actor" />
-            </SummaryCard>
-          </Column>
-          <Column>
-            <SummaryCard {...getLayoutProps()}>
-              <SummaryCardHeader title="Threat report" />
-            </SummaryCard>
-          </Column>
-          <Column>
-            <SummaryCard {...getLayoutProps()}>
-              <SummaryCardHeader title="IP report" />
-            </SummaryCard>
-          </Column>
-          <Column>
-            <SummaryCard {...getLayoutProps()}>
-              <SummaryCardHeader title="Vulnerability report" />
-            </SummaryCard>
-          </Column>
-        </Row>
-      )}
-    </CardModule>
-
-    <Row>
-      <ColumnWithBackground>
-        <ICAModule>
-          {({ getLayoutProps }) => (
-            <>
-              <TitleBarModule element="h4" title="Indicators" />
-
-              <Row>
-                <Column {...getLayoutProps({ lg: 3, md: 2, sm: 2 })}>
-                  <ICA label="Malware" value={11} />
-                </Column>
-                <Column {...getLayoutProps({ lg: 3, md: 2, sm: 2 })}>
-                  <ICA label="IPs" value={8} />
-                </Column>
-                <Column {...getLayoutProps({ lg: 3, md: 2, sm: 2 })}>
-                  <ICA label="URLs" value={9} />
-                </Column>
-                <Column {...getLayoutProps({ lg: 3, md: 2, sm: 2 })}>
-                  <ICA label="VULs" value={1} />
-                </Column>
-              </Row>
-            </>
-          )}
-        </ICAModule>
-
-        <Row condensed>
-          <Column>
-            <DataTablePagination
-              headers={[
-                {
-                  header: 'Name',
-                  key: 'name',
-                },
-                {
-                  header: 'Last sighted',
-                  key: 'lastSighted',
-                },
-              ]}
-              pageSize={5}
-              pageSizes={[5, 10, 25, 50]}
-              rows={[
-                {
-                  id: '0',
-                  lastSighted: 'Feb 3 2019 12:00 PM EST',
-                  name: (
-                    <Decorator
-                      score={7}
-                      type="MAL"
-                      value="5020c08bcc061236643293bf0d897321"
-                    />
-                  ),
-                },
-                {
-                  id: '1',
-                  lastSighted: 'Feb 5 2019 12:00 PM EST',
-                  name: (
-                    <Decorator
-                      score={7}
-                      type="MAL"
-                      value="aca7037286b64b0da05c9708d647c013"
-                    />
-                  ),
-                },
-                {
-                  id: '2',
-                  lastSighted: 'Feb 7 2019 12:00 PM EST',
-                  name: (
-                    <Decorator
-                      score={7}
-                      type="MAL"
-                      value="bd9e4c82bf12c4e7a58221fc52fed705"
-                    />
-                  ),
-                },
-                {
-                  id: '3',
-                  lastSighted: 'Apr 1 2019 12:00 PM EST',
-                  name: (
-                    <Decorator score={0} type="IP" value="103.243.175.181" />
-                  ),
-                },
-                {
-                  id: '4',
-                  lastSighted: 'Apr 1 2019 12:00 PM EST',
-                  name: (
-                    <Decorator
-                      score={7}
-                      type="MAL"
-                      value="5020c08bcc061236643293bf0d897321"
-                    />
-                  ),
-                },
-              ]}
-            />
-          </Column>
-        </Row>
-      </ColumnWithBackground>
-    </Row>
-  </>
-);
-
-export const Overview = () => (
+export const Default = () => (
   <>
     <ActionBarModule>
       <Tag type="gray">Closed</Tag>
